Rename module to flapison
