+++ flapison/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from flapison.api import Api
+from flapison.resource import ResourceList, ResourceDetail, ResourceRelationship
+from flapison.exceptions import JsonApiException
+
+__all__ = [
+    'Api',
+    'ResourceList',
+    'ResourceDetail',
+    'ResourceRelationship',
+    'JsonApiException'
+]
