+++ flapison/api.py
@@ -10,8 +10,8 @@ from functools import wraps
 from flask import abort
 from flask import request
 
-from flask_rest_jsonapi.decorators import jsonapi_exception_formatter
-from flask_rest_jsonapi.resource import ResourceList, ResourceRelationship
+from flapison.decorators import jsonapi_exception_formatter
+from flapison.resource import ResourceList, ResourceRelationship
 
 
 class Api(object):
@@ -67,7 +67,7 @@ class Api(object):
     def route(self, resource, view, *urls, **kwargs):
         """Create an api view.
 
-        :param Resource resource: a resource class inherited from flask_rest_jsonapi.resource.Resource
+        :param Resource resource: a resource class inherited from flapison.resource.Resource
         :param str view: the view name
         :param list urls: the urls of the view
         :param dict kwargs: additional options of the route
