--- flask_rest_jsonapi/__init__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from flask_rest_jsonapi.api import Api
-from flask_rest_jsonapi.resource import ResourceList, ResourceDetail, ResourceRelationship
-from flask_rest_jsonapi.exceptions import JsonApiException
-
-__all__ = [
-    'Api',
-    'ResourceList',
-    'ResourceDetail',
-    'ResourceRelationship',
-    'JsonApiException'
-]
