+++ tests/test_sqlalchemy_data_layer.py
@@ -15,16 +15,16 @@ from sqlalchemy import create_engine, Column, Integer, DateTime, String, Foreign
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
 
-import flask_rest_jsonapi.decorators
-import flask_rest_jsonapi.resource
-import flask_rest_jsonapi.schema
-from flask_rest_jsonapi import Api, ResourceList, ResourceDetail, ResourceRelationship, JsonApiException
-from flask_rest_jsonapi.data_layers.alchemy import SqlalchemyDataLayer
-from flask_rest_jsonapi.data_layers.base import BaseDataLayer
-from flask_rest_jsonapi.data_layers.filtering.alchemy import Node
-from flask_rest_jsonapi.exceptions import RelationNotFound, InvalidSort, InvalidFilters, InvalidInclude, BadRequest
-from flask_rest_jsonapi.pagination import add_pagination_links
-from flask_rest_jsonapi.querystring import QueryStringManager as QSManager
+import flapison.decorators
+import flapison.resource
+import flapison.schema
+from flapison import Api, ResourceList, ResourceDetail, ResourceRelationship, JsonApiException
+from flapison.data_layers.alchemy import SqlalchemyDataLayer
+from flapison.data_layers.base import BaseDataLayer
+from flapison.data_layers.filtering.alchemy import Node
+from flapison.exceptions import RelationNotFound, InvalidSort, InvalidFilters, InvalidInclude, BadRequest
+from flapison.pagination import add_pagination_links
+from flapison.querystring import QueryStringManager as QSManager
 
 
 def test_add_pagination_links(app):
@@ -74,9 +74,9 @@ def test_Node(person_model, person_schema, monkeypatch):
 def test_check_method_requirements(monkeypatch):
     self = type('self', (object,), dict())
     request = type('request', (object,), dict(method='GET'))
-    monkeypatch.setattr(flask_rest_jsonapi.decorators, 'request', request)
+    monkeypatch.setattr(flapison.decorators, 'request', request)
     with pytest.raises(Exception):
-        flask_rest_jsonapi.decorators.check_method_requirements(lambda: 1)(self())
+        flapison.decorators.check_method_requirements(lambda: 1)(self())
 
 
 def test_json_api_exception():
@@ -113,11 +113,11 @@ def test_resource(app, person_model, person_schema, session, monkeypatch):
         rl.schema = person_schema
         rd._data_layer = dl
         rd.schema = person_schema
-        monkeypatch.setattr(flask_rest_jsonapi.resource, 'request', request)
-        monkeypatch.setattr(flask_rest_jsonapi.decorators, 'current_app', app)
-        monkeypatch.setattr(flask_rest_jsonapi.decorators, 'request', request)
+        monkeypatch.setattr(flapison.resource, 'request', request)
+        monkeypatch.setattr(flapison.decorators, 'current_app', app)
+        monkeypatch.setattr(flapison.decorators, 'request', request)
         monkeypatch.setattr(rl.schema, 'load', schema_load_mock)
-        r = super(flask_rest_jsonapi.resource.Resource, ResourceList) \
+        r = super(flapison.resource.Resource, ResourceList) \
             .__new__(ResourceList)
         with pytest.raises(Exception):
             r.dispatch_request()
@@ -129,16 +129,16 @@ def test_compute_schema(person_schema):
     query_string = {'page[number]': '3', 'fields[person]': list()}
     qsm = QSManager(query_string, person_schema)
     with pytest.raises(InvalidInclude):
-        flask_rest_jsonapi.schema.compute_schema(person_schema, dict(), qsm, ['id'])
-    flask_rest_jsonapi.schema.compute_schema(person_schema, dict(only=list()), qsm, list())
+        flapison.schema.compute_schema(person_schema, dict(), qsm, ['id'])
+    flapison.schema.compute_schema(person_schema, dict(only=list()), qsm, list())
 
 
 def test_compute_schema_propagate_context(person_schema, computer_schema):
     query_string = {}
     qsm = QSManager(query_string, person_schema)
-    schema = flask_rest_jsonapi.schema.compute_schema(person_schema, dict(), qsm, ['computers'])
+    schema = flapison.schema.compute_schema(person_schema, dict(), qsm, ['computers'])
     assert schema.declared_fields['computers'].__dict__['_Relationship__schema'].__dict__['context'] == dict()
-    schema = flask_rest_jsonapi.schema.compute_schema(person_schema, dict(context=dict(foo='bar')), qsm, ['computers'])
+    schema = flapison.schema.compute_schema(person_schema, dict(context=dict(foo='bar')), qsm, ['computers'])
     assert schema.declared_fields['computers'].__dict__['_Relationship__schema'].__dict__['context'] == dict(foo='bar')
